fix: resolve all SwiftLint violations
- Extend swiftlint:disable block in View.swift to cover all buildBlock
  overloads (large_tuple, function_parameter_count)
- Extract segmented control code from ViewBuilder+Controls.swift into
  ViewBuilder+SegmentedControl.swift with focused helper functions
- Split segmented control tests into SegmentedControlTests.swift to
  resolve file_length and type_body_length violations
- Auto-fix trailing comma violations across test and example files
