+++ Sources/Whisker/Core/ViewBuilder+Controls.swift
@@ -233,208 +233,4 @@ extension NodeViewBuilder {
         }
     }
 
-    // MARK: - Segmented Control
-
-    private func buildSegmentedControlNode(_ node: Node, segmented: SegmentedControl) {
-        node.isFocusable = true
-
-        node[.keyHandler] = { (event: KeyEvent) in
-            let options = segmented.options
-            guard !options.isEmpty else { return }
-
-            let current = min(max(segmented.selection.wrappedValue, 0), options.count - 1)
-            var next = current
-
-            switch event.key {
-            case .left:
-                if segmented.wraps && current == 0 {
-                    next = options.count - 1
-                } else {
-                    next = max(0, current - 1)
-                }
-            case .right:
-                if segmented.wraps && current == options.count - 1 {
-                    next = 0
-                } else {
-                    next = min(options.count - 1, current + 1)
-                }
-            default:
-                return
-            }
-
-            if next != segmented.selection.wrappedValue {
-                segmented.selection.wrappedValue = next
-                Application.shared?.scheduleUpdate()
-            }
-        }
-
-        node.render = { [weak node] frame, buffer in
-            guard let node = node else { return }
-            let options = segmented.options
-            guard !options.isEmpty else { return }
-
-            let selectedIndex = min(max(segmented.selection.wrappedValue, 0), options.count - 1)
-            let baseStyle = Style().resolved(with: node.environment)
-
-            switch segmented.overflow {
-            case .scroll:
-                // Pre-compute segment start positions in content space
-                var segmentStarts: [Int] = []
-                var cx = 0
-                for (index, option) in options.enumerated() {
-                    segmentStarts.append(cx)
-                    cx += option.count + 2
-                    if index < options.count - 1 {
-                        cx += 1
-                    }
-                }
-                let contentWidth = cx
-
-                // Compute scroll offset to keep selected segment visible
-                var scrollOffset = 0
-                let hasOverflow = contentWidth > frame.width
-                // Reserve 1 char at each edge for scroll indicators
-                let indicatorWidth = hasOverflow ? 1 : 0
-                let viewportWidth = frame.width - (indicatorWidth * 2)
-
-                if hasOverflow && viewportWidth > 0 {
-                    let selStart = segmentStarts[selectedIndex]
-                    let selEnd = selStart + options[selectedIndex].count + 2
-
-                    if selEnd > scrollOffset + viewportWidth {
-                        scrollOffset = selEnd - viewportWidth
-                    }
-                    if selStart < scrollOffset {
-                        scrollOffset = selStart
-                    }
-                    scrollOffset = min(scrollOffset, max(0, contentWidth - viewportWidth))
-                    scrollOffset = max(scrollOffset, 0)
-                }
-
-                // Draw left scroll indicator
-                let showLeftIndicator = hasOverflow && scrollOffset > 0
-                let showRightIndicator = hasOverflow && scrollOffset + viewportWidth < contentWidth
-                let dimStyle = Style(foreground: .brightBlack)
-
-                if hasOverflow {
-                    let leftChar: Character = showLeftIndicator ? "\u{2039}" : " "
-                    buffer.draw(leftChar, at: Position(x: frame.x, y: frame.y), style: dimStyle)
-                }
-
-                // Render segments in the viewport area
-                let viewportX = frame.x + indicatorWidth
-                var cursorX = viewportX - scrollOffset
-
-                for (index, option) in options.enumerated() {
-                    let segmentText = " \(option) "
-                    var style = baseStyle
-                    if index == selectedIndex {
-                        style.attributes.insert(.reverse)
-                    }
-                    if index == selectedIndex && node.isFocused {
-                        style = style.bold()
-                    }
-
-                    for char in segmentText {
-                        if cursorX >= viewportX + viewportWidth { break }
-                        if cursorX >= viewportX {
-                            buffer.draw(char, at: Position(x: cursorX, y: frame.y), style: style)
-                        }
-                        cursorX += 1
-                    }
-
-                    if index < options.count - 1 {
-                        if cursorX >= viewportX + viewportWidth { break }
-                        if cursorX >= viewportX {
-                            buffer.draw(" ", at: Position(x: cursorX, y: frame.y), style: .default)
-                        }
-                        cursorX += 1
-                    }
-                }
-
-                // Draw right scroll indicator
-                if hasOverflow {
-                    let rightChar: Character = showRightIndicator ? "\u{203A}" : " "
-                    buffer.draw(
-                        rightChar, at: Position(x: frame.x + frame.width - 1, y: frame.y),
-                        style: dimStyle)
-                }
-
-            case .wrap:
-                let availableWidth = frame.width
-                guard availableWidth > 0 else { return }
-
-                // Lay out segments into lines
-                var cursorX = frame.x
-                var cursorY = frame.y
-
-                for (index, option) in options.enumerated() {
-                    let segmentWidth = option.count + 2
-
-                    // Wrap to next line if this segment won't fit
-                    if cursorX > frame.x && cursorX - frame.x + segmentWidth > availableWidth {
-                        cursorX = frame.x
-                        cursorY += 1
-                    }
-
-                    if cursorY >= frame.y + frame.height { break }
-
-                    let segmentText = " \(option) "
-                    var style = baseStyle
-                    if index == selectedIndex {
-                        style.attributes.insert(.reverse)
-                    }
-                    if index == selectedIndex && node.isFocused {
-                        style = style.bold()
-                    }
-
-                    for char in segmentText {
-                        if cursorX < frame.x + availableWidth {
-                            buffer.draw(char, at: Position(x: cursorX, y: cursorY), style: style)
-                        }
-                        cursorX += 1
-                    }
-
-                    if index < options.count - 1 && cursorX < frame.x + availableWidth {
-                        buffer.draw(" ", at: Position(x: cursorX, y: cursorY), style: .default)
-                        cursorX += 1
-                    }
-                }
-            }
-        }
-
-        node.layout = { proposal, _ in
-            let options = segmented.options
-            guard !options.isEmpty else { return (Size(width: 0, height: 1), []) }
-            let spacing = options.count > 1 ? (options.count - 1) : 0
-            let contentWidth = options.reduce(0) { $0 + $1.count + 2 } + spacing
-
-            switch segmented.overflow {
-            case .scroll:
-                let width = proposal.width.resolve(with: contentWidth)
-                return (Size(width: width, height: 1), [])
-
-            case .wrap:
-                let availableWidth = proposal.width.resolve(with: contentWidth)
-                guard availableWidth > 0 else { return (Size(width: 0, height: 1), []) }
-
-                // Calculate number of lines needed
-                var lineCount = 1
-                var x = 0
-                for (index, option) in options.enumerated() {
-                    let segmentWidth = option.count + 2
-                    let spacingWidth = index < options.count - 1 ? 1 : 0
-
-                    if x > 0 && x + segmentWidth > availableWidth {
-                        lineCount += 1
-                        x = segmentWidth + spacingWidth
-                    } else {
-                        x += segmentWidth + spacingWidth
-                    }
-                }
-
-                return (Size(width: availableWidth, height: lineCount), [])
-            }
-        }
-    }
 }
